+++ src/main/java/com/baomidou/dynamic/datasource/creator/BasicDataSourceCreator.java
@@ -1,0 +1,89 @@
+/**
+ * Copyright © 2018 organization baomidou
+ * <pre>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * <pre/>
+ */
+package com.baomidou.dynamic.datasource.creator;
+
+import com.baomidou.dynamic.datasource.exception.ErrorCreateDataSourceException;
+import com.baomidou.dynamic.datasource.spring.boot.autoconfigure.DataSourceProperty;
+import java.lang.reflect.Method;
+import javax.sql.DataSource;
+import lombok.Data;
+
+/**
+ * @author TaoYu
+ * @date 2020/1/21
+ */
+@Data
+public class BasicDataSourceCreator {
+
+  private static Method createMethod;
+  private static Method typeMethod;
+  private static Method urlMethod;
+  private static Method usernameMethod;
+  private static Method passwordMethod;
+  private static Method driverClassNameMethod;
+  private static Method buildMethod;
+
+  static {
+    //to support springboot 1.5 and 2.x
+    Class<?> builderClass = null;
+    try {
+      builderClass = Class.forName("org.springframework.boot.jdbc.DataSourceBuilder");
+    } catch (Exception ignored) {
+    }
+    if (builderClass == null) {
+      try {
+        builderClass = Class.forName("org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder");
+      } catch (Exception e) {
+      }
+    }
+    if (builderClass != null) {
+      try {
+        createMethod = builderClass.getDeclaredMethod("create");
+        typeMethod = builderClass.getDeclaredMethod("type", Class.class);
+        urlMethod = builderClass.getDeclaredMethod("url", String.class);
+        usernameMethod = builderClass.getDeclaredMethod("username", String.class);
+        passwordMethod = builderClass.getDeclaredMethod("password", String.class);
+        driverClassNameMethod = builderClass.getDeclaredMethod("driverClassName", String.class);
+        buildMethod = builderClass.getDeclaredMethod("build");
+      } catch (Exception e) {
+        e.printStackTrace();
+      }
+    }
+  }
+
+  /**
+   * 创建基础数据源
+   *
+   * @param dataSourceProperty 数据源参数
+   * @return 数据源
+   */
+  public DataSource createDataSource(DataSourceProperty dataSourceProperty) {
+    try {
+      Object o1 = createMethod.invoke(null);
+      Object o2 = typeMethod.invoke(o1, dataSourceProperty.getType());
+      Object o3 = urlMethod.invoke(o2, dataSourceProperty.getUrl());
+      Object o4 = usernameMethod.invoke(o3, dataSourceProperty.getUsername());
+      Object o5 = passwordMethod.invoke(o4, dataSourceProperty.getPassword());
+      Object o6 = driverClassNameMethod.invoke(o5, dataSourceProperty.getDriverClassName());
+      return (DataSource) buildMethod.invoke(o6);
+    } catch (Exception e) {
+      throw new ErrorCreateDataSourceException(
+          "dynamic-datasource create basic database named " + dataSourceProperty.getPollName() + " error");
+    }
+  }
+
+}
